Fix button styles
I thought that this was unnecessary becouse this new class was already stated in assets/css/socials-btn-styles.css, but that seems to not be loading at all. Will leave it for now and figure out why is it not working then delete either this change or one in assets/css and theme.yml.